+++ src/main/java/esn/entities/secondary/GenChatMessage.java
@@ -1,11 +1,10 @@
 package esn.entities.secondary;
 
-import esn.db.UserDAO;
 import esn.entities.User;
 
 public class GenChatMessage extends AbstractMessage{
 
-    public GenChatMessage(int id, String text, int orgId, User user) {
+    public GenChatMessage(long id, String text, int orgId, User user) {
         super(id, text, orgId, user);
     }
